requirements.txt
django==4.2.0
django-mirage-field
django-geoposition-2
mysqlclient
django_q
croniter
ansible==6.0.0
djangorestframework==3.14.0
django-filter
django-redis
django-environ
paramiko>=3.2.0
drf-jwt
django-cors-headers
django-mptt
